Moved slice for stations/channels
