+++ apps/client/src/app/reducers.ts
@@ -2,8 +2,8 @@ import { combineReducers } from 'redux'
 import authentication from 'features/Me/Authentication/slice'
 import me from 'features/Me/slice'
 import logout from 'features/Me/Logout/slice'
-import stations from 'features/Stations/List/slice'
-import channels from 'features/Channels/List/slice'
+import stations from 'features/Stations/slice'
+import channels from 'features/Channels/slice'
 import messages from 'features/Messages/List/slice'
 import messenger from 'features/Messages/Messenger/slice'
